requirements.txt
asgiref==3.4.1
bitrix24-rest==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
idna==3.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
sentry-sdk==1.5.5
psycopg2-binary==2.9.3
gunicorn==20.0.4